requirements_colab.txt
# Requirements for Google Colab
# Install with: pip install -q -r requirements_colab.txt

stable-baselines3[extra]>=2.0.0
gymnasium>=0.28.0
geopandas>=0.13.0
osmnx>=1.5.0
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
Pillow>=10.0.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
rtree>=1.0.0

# PyTorch is pre-installed in Colab, but if needed:
# torch>=2.0.0

